While importing a new profile if profile cannot be imported then that profile will not be shown in drop down menu.

+++ Obi/Obi/Dialogs/Preferences.cs
@@ -41,6 +41,7 @@ namespace Obi.Dialogs
       //  private bool m_FlagComboBoxIndexChange = false;
        // private int m_IndexOfLevelCombox = 0;
         private bool m_UpdateBackgroundColorRequired = false;
+        private bool m_IsProfileLoadFailed = false;
         private int m_Audio_CleanupMaxFileSizeInMB = 100;
         private List<string> m_FontPermissible; //@fontconfig
       //  private bool m_IsComboBoxExpanded = false;
@@ -2809,7 +2810,11 @@ m_cb_ChooseFont.Visible = false;
             if (profilePath != null && System.IO.File.Exists(profilePath))
             {
                 Settings saveProfile = Settings.GetSettingsFromSavedProfile(profilePath,IsImport);
-                if (saveProfile == null) return;
+                if (saveProfile == null)
+                {
+                    m_IsProfileLoadFailed = true;
+                    return;
+                }
                 string profileName = "";
                 if (m_cb_SelectProfile.SelectedIndex >= 0 && m_cb_SelectProfile.SelectedIndex < m_cb_SelectProfile.Items.Count)
                 {
@@ -3124,11 +3129,11 @@ m_cb_ChooseFont.Visible = false;
                     if (!System.IO.Directory.Exists(customProfilesDirectory)) System.IO.Directory.CreateDirectory(customProfilesDirectory);
                     string newCustomFilePath = System.IO.Path.Combine(customProfilesDirectory,
                         System.IO.Path.GetFileName(fileDialog.FileName));
-                    if (!File.Exists(newCustomFilePath))
+                    if (!File.Exists(newCustomFilePath) && !m_IsProfileLoadFailed)
                     {
                         System.IO.File.Copy(fileDialog.FileName, newCustomFilePath, true);
                     }
-                    if (!m_cb_SelectProfile.Items.Contains(System.IO.Path.GetFileNameWithoutExtension(newCustomFilePath)))
+                    if (!m_cb_SelectProfile.Items.Contains(System.IO.Path.GetFileNameWithoutExtension(newCustomFilePath)) && !m_IsProfileLoadFailed)
                     {
                         m_cb_SelectProfile.Items.Add(System.IO.Path.GetFileNameWithoutExtension(newCustomFilePath));
                         m_cb_Profile1.Items.Add(System.IO.Path.GetFileNameWithoutExtension(newCustomFilePath));
